# Precompute offsets rather than using a shared memory buffer for conditional joins