… cpp/tests/join/conditional_join_tests.cu