@vyasr 2022-01-03
@jrhemstad let me know what you think of this change. It makes the API a little more complicated and increases memory usage, but conversely it improves performance by ~10% and simplifies the internals substantially by removing the need for the shared memory logic.

<details><summary>Benchmarks</summary>
<p>

Before:
```
----------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                    Time             CPU   Iterations
----------------------------------------------------------------------------------------------------------------------------------------------
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/100000/manual_time                   818 ms          820 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/400000/manual_time                  2678 ms         2678 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/400000/100000/manual_time                  2688 ms         2688 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/1000000/manual_time                 6660 ms         6660 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/100000/manual_time                   762 ms          762 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/400000/manual_time                  2862 ms         2862 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/400000/100000/manual_time                  2869 ms         2869 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/1000000/manual_time                 7101 ms         7100 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/100000/manual_time            1474 ms         1474 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/400000/manual_time            5620 ms         5620 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/400000/100000/manual_time            5775 ms         5775 ms            1
```

After:

```
----------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                    Time             CPU   Iterations
----------------------------------------------------------------------------------------------------------------------------------------------
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/100000/manual_time                   750 ms          752 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/400000/manual_time                  2390 ms         2390 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/400000/100000/manual_time                  2401 ms         2401 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/1000000/manual_time                 5919 ms         5919 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/100000/manual_time                   671 ms          671 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/400000/manual_time                  2560 ms         2560 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/400000/100000/manual_time                  2561 ms         2561 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/1000000/manual_time                 6319 ms         6319 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/100000/manual_time            1367 ms         1367 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/400000/manual_time            5236 ms         5236 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/400000/100000/manual_time            5371 ms         5370 ms            1
```
</p>
</details>