@vyasr 2022-07-13
Before proceeding any further, @jrhemstad wanted to see what performance looks like if we implement the same change in `cuco::static_multimap::retrieve`. That would involve [removing the `output_buffer`](https://github.com/NVIDIA/cuCollections/blob/dev/include/cuco/detail/static_multimap/kernels.cuh#L380) and instead launching two separate kernels. I'm going to close this PR for now, it can be reopened if we decide to proceed.

CC @PointKernel in case you get a chance to test out the perf impacts of this in cuco at some point before I do.